+++ tests/testthat/test-run-model.R
@@ -1,38 +1,5 @@
 context("run-model")
 
-## A single set of valid model options and data, update once instead of copying
-## for every test.
-
-a_hintr_data <- list(
-  pjnz = system_file("extdata/mwi2019.PJNZ"),
-  population = system_file("extdata/population/population_agesex.csv"),
-  shape = system_file("extdata/areas/area_merged.geojson"),
-  survey = system_file("extdata/survey/survey_hiv_indicators.csv"),
-  art_number = system_file("extdata/programme/art_number.csv"),
-  anc_testing = system_file("extdata/programme/anc_testing.csv")
-)
-
-a_hintr_options <- list(
-  area_scope = "MWI_1_1",
-  area_level = 4,
-  calendar_quarter_t1 = "CY2016Q1",
-  calendar_quarter_t2 = "CY2018Q3",
-  survey_prevalence = c("MWI2016PHIA", "MWI2015DHS"),
-  survey_art_coverage = "MWI2016PHIA",
-  survey_recently_infected = "MWI2016PHIA",
-  include_art_t1 = "true",
-  include_art_t2 = "true",
-  anc_prevalence_year1 = 2016,
-  anc_prevalence_year2 = 2018,
-  anc_art_coverage_year1 = 2016,
-  anc_art_coverage_year2 = 2018,
-  artattend = FALSE,
-  rng_seed = 17,
-  no_of_samples = 20,
-  max_iter = 250,
-  permissive = FALSE
-)
-
 test_that("model can be run", {
 
   output_path <- tempfile()
@@ -53,7 +20,7 @@ test_that("model can be run", {
                  "calendar_quarter", "quarter_id", "quarter_label",
                  "indicator", "indicator_id", "indicator_label",
                  "mean", "se", "median", "mode", "lower", "upper"))
-  expect_true(nrow(output) == 22320)
+  expect_true(nrow(output) == 30624)
   expect_equal(model_run$spectrum_path, output_spectrum)
   file_list <- unzip(model_run$spectrum_path, list = TRUE)
   ## Note that this test is likely quite platform specific
@@ -63,7 +30,10 @@ test_that("model can be run", {
     file_list$Name,
     c("boundaries.geojson", "indicators.csv", "meta_age_group.csv",
       "meta_area.csv", "meta_indicator.csv", "meta_period.csv",
-      "info/", info_names))
+      "info/", info_names,
+      "fit/", "fit/spectrum_calibration.csv", "fit/calibration_options.csv")
+  )
+      
 
   ## TODO: replace with checks for spectrum digest once function to create
   ## that has been added mrc-636
@@ -73,7 +43,9 @@ test_that("model can be run", {
     file_list$Name,
     c("boundaries.geojson", "indicators.csv", "meta_age_group.csv",
       "meta_area.csv", "meta_indicator.csv", "meta_period.csv",
-      "info/", info_names))
+      "info/", info_names,
+      "fit/", "fit/spectrum_calibration.csv", "fit/calibration_options.csv")
+  )
 
   tmp <- tempfile()
   unzip(model_run$spectrum_path, exdir = tmp, files = info_names)
@@ -103,6 +75,7 @@ test_that("model can be run without programme data", {
   options$anc_prevalence_year2 <- NULL
   options$anc_art_coverage_year1 <- NULL
   options$anc_art_coverage_year2 <- NULL
+  options$artattend <- "false"
 
   output_path <- tempfile()
   output_spectrum <- tempfile(fileext = ".zip")
@@ -119,7 +92,7 @@ test_that("model can be run without programme data", {
                  "calendar_quarter", "quarter_id", "quarter_label",
                  "indicator", "indicator_id", "indicator_label",
                  "mean", "se", "median", "mode", "lower", "upper"))
-  expect_true(nrow(output) == 22320)
+  expect_true(nrow(output) == 30624)
 
   expect_equal(model_run$spectrum_path, output_spectrum)
   file_list <- unzip(model_run$spectrum_path, list = TRUE)
@@ -130,7 +103,9 @@ test_that("model can be run without programme data", {
     file_list$Name,
     c("boundaries.geojson", "indicators.csv", "meta_age_group.csv",
       "meta_area.csv", "meta_indicator.csv", "meta_period.csv",
-      "info/", info_names))
+      "info/", info_names,
+      "fit/", "fit/spectrum_calibration.csv", "fit/calibration_options.csv")
+  )
 
   ## TODO: replace with checks for spectrum digest once function to create
   ## that has been added mrc-636
@@ -140,9 +115,41 @@ test_that("model can be run without programme data", {
     file_list$Name,
     c("boundaries.geojson", "indicators.csv", "meta_age_group.csv",
       "meta_area.csv", "meta_indicator.csv", "meta_period.csv",
-      "info/", info_names))
+      "info/", info_names,
+      "fit/", "fit/spectrum_calibration.csv", "fit/calibration_options.csv")
+  )
 
 })
+
+test_that("model fit without survey ART and survey recency data", {
+
+  ## !!! TODO: need to get this working or validation flags
+  skip("Need to return to either get working or set validation flags")
+  
+  options <- a_hintr_options
+  options$survey_art_coverage <- NULL
+  expect_error(
+    hintr_run_model(a_hintr_data, options, tempfile(), tempfile(), tempfile()),
+    NA)
+
+  options <- a_hintr_options
+  options$survey_recently_infected <- NULL
+  expect_error(
+    hintr_run_model(a_hintr_data, options, tempfile(), tempfile(), tempfile()),
+    NA)
+
+  ## No survey ART coverage or ART programme data
+  ## !!! TODO: This is a situation that **should** be supported, needs attention
+  options <- a_hintr_options
+  options$survey_art_coverage <- NULL
+  options$include_art_t1 = "false"
+  options$include_art_t2 = "false"
+  expect_error(
+    hintr_run_model(a_hintr_data, options, tempfile(), tempfile(), tempfile()),
+    "false convergence \\(8\\)")
+  
+}
+)
 
 test_that("progress messages are printed", {
   skip_on_covr()
@@ -182,26 +189,13 @@ test_that("progress messages are printed", {
 })
 
 test_that("model run throws error for invalid inputs", {
-  options_bad <- list(
-    area_scope = "MWI",
-    calendar_quarter_t1 = "CY2016Q1",
-    survey_prevalence = c("MWI2016PHIA", "MWI2015DHS"),
-    survey_art_coverage = "MWI2016PHIA",
-    survey_recently_infected = "MWI2016PHIA",
-    include_art_t1 = "true",
-    include_art_t2 = "true",
-    anc_prevalence_year1 = 2016,
-    anc_prevalence_year2 = 2018,
-    anc_art_coverage_year1 = 2016,
-    anc_art_coverage_year2 = 2018,
-    no_of_samples = 20
-  )
   output_path <- tempfile()
   output_spectrum <- tempfile(fileext = ".zip")
   summary_path <- tempfile(fileext = ".zip")
   expect_error(
-    hintr_run_model(data, options_bad, output_path, output_spectrum,
-                    summary_path)
+    hintr_run_model(data, a_hintr_options_bad,
+                    output_path, output_spectrum, summary_path)
+                    
   )
 })
 
@@ -212,11 +206,14 @@ test_that("setting rng_seed returns same output", {
   
   options <- a_hintr_options
   options$survey_prevalence = "MWI2016PHIA"
-  options$survey_art_coverage <- NULL
+  options$survey_art_coverage <- "MWI2016PHIA"
   options$survey_recently_infected <- NULL
   options$include_art_t1 = "false"
   options$include_art_t2 = "false"
-
+  options$artattend <- "false"
+  options$spectrum_plhiv_calibration_level <- "none"
+  options$spectrum_artnum_calibration_level <- "none"
+  
   output_path <- tempfile()
   output_spectrum <- tempfile(fileext = ".zip")
   summary_path <- tempfile(fileext = ".zip")
@@ -266,6 +263,7 @@ test_that("exceeding max_iterations convergence error or warning", {
   options$survey_recently_infected <- NULL
   options$include_art_t1 = "false"
   options$include_art_t2 = "false"
+  options$artattend <- "false"
   options$max_iterations <- 5
 
   output_path <- tempfile()
@@ -276,11 +274,21 @@ test_that("exceeding max_iterations convergence error or warning", {
                                output_path, output_spectrum,
                                summary_path))
 
-  options$permissive <- TRUE
+  options$permissive <- "true"
   output_path <- tempfile()
   output_spectrum <- tempfile(fileext = ".zip")
   summary_path <- tempfile(fileext = ".zip")
   expect_warning(hintr_run_model(data, options,
                                  output_path, output_spectrum,
                                  summary_path))
+})
+
+
+test_that("naomi_info_input(data) handles NULL string", {
+
+  data <- list(file1 = "file1.ext",
+               file2 = "file2.ext",
+               file3 = NULL)
+  
+  expect_equal(nrow(naomi_info_input(data)), 3)
 })